Fixed bug where level color would change in the editor

+++ Assets/Scripts/LevelBuilder.cs
@@ -219,8 +219,8 @@ public class LevelBuilder : MonoBehaviour {
 	}
 
 	void setCamera() {
-		float yHeight1 = Mathf.Abs(xMin - xMax) + ((Screen.height/Screen.width) * 0.5f);
-		float yHeight2 = Mathf.Abs(zMin - zMax) + ((Screen.height/Screen.width) * 0.5f);
+		float yHeight1 = Mathf.Abs(xMin - xMax) + (((float)Screen.height / (float)Screen.width) * 0.5f);
+		float yHeight2 = Mathf.Abs(zMin - zMax) + (((float)Screen.height / (float)Screen.width) * 0.5f);
 		Camera.main.orthographicSize = ((rightMost - leftMost) - (rightMost - leftMost) / 4.5f) + ((float)Screen.height / (float)Screen.width);
 		Camera.main.orthographicSize = Mathf.Clamp(Camera.main.orthographicSize, startOrthoMin, startOrthoMax + ((float)Screen.height / (float)Screen.width));
 		if ((yHeight2 / yHeight1) < 1.625f) {
